style(msvc): fix C4244/C4267 in every remaining 2-site file (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; [NotesLoader],[NoteData]
-tagged tests (646/46) re-checked identical.

26 files, 52 sites: StepsUtil.cpp, StageStats.cpp, Sprite.cpp,
ScrollBar.cpp, ScreenSelectMusic.cpp, ScreenOptionsManageProfiles.cpp,
ScreenMapControllers.cpp, ScreenInstallOverlay.cpp, SampleHistory.cpp,
RandomSample.cpp, RageUtil_CharConversions.cpp,
RageSoundReader_SpeedChange.cpp, RageSoundReader_Resample_Good.cpp,
RageSoundReader_Chain.cpp, RageModelGeometry.cpp,
RageFileDriverTimeout.cpp, ProfileManager.cpp, PlayerOptions.cpp,
Player.cpp, NotesLoaderSMA.cpp (SS5-protected, characterization-only),
NotesLoaderJson.cpp (SS5-protected, characterization-only),
NoteData.cpp (SS5-protected NoteData*, characterization-only),
MsdFile.h, GameSoundManager.cpp, CsvFile.cpp, CharacterManager.cpp.

Same recurring narrowing patterns throughout: size_t/double-family
values narrowed at wrap()/RandomInt() calls, iterator-diffs,
reverse-loop bounds, RString::Left/Right length args, and Lua
lua_push*/lua_raw* bindings.

+++ src/CsvFile.cpp
@@ -72,7 +72,7 @@ bool CsvFile::ReadFile( RageFileBasic &f )
 				while(true);
 
 				RString sValue = line;
-				sValue = sValue.Left( iEnd );
+				sValue = sValue.Left( static_cast<int>(iEnd) );
 				vs.push_back( sValue );
 
 				line.erase( line.begin(), line.begin()+iEnd );
@@ -87,7 +87,7 @@ bool CsvFile::ReadFile( RageFileBasic &f )
 					iEnd = line.size();	// didn't find an end.  Take the whole line
 
 				RString sValue = line;
-				sValue = sValue.Left( iEnd );
+				sValue = sValue.Left( static_cast<int>(iEnd) );
 				vs.push_back( sValue );
 
 				line.erase( line.begin(), line.begin()+iEnd );
